Add a 'download data' page to the admin site.
