+++ project/templates/admin/justfix/download_data.html
@@ -1,0 +1,17 @@
+{% extends "admin/base_site.html" %}
+{% block content %}
+<ul>
+    <li><a href="#TODO">User statistics (CSV)</a> - Anonymized statistics about each user,
+        including when they completed onboarding, sent a letter of complaint,
+        and so on.
+    </li>
+    <li><a href="#TODO">User statistics with PII (CSV)</a> -
+        This is like the user statistics CSV but also includes the BBL of each user,
+        which could potentially be used to personally identify them.
+    </li>
+</ul>
+<p>
+    Please be careful when downloading data that contains personally-identifiable
+    information (PII).  We don't want that information falling into the wrong hands!
+</p>
+{% endblock %}
